# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "additionalProperties": false,
        "properties": {
          "brain": {
            "$ref": "#/components/schemas/BrainActivity"
          },
          "classification": {
            "$ref": "#/components/schemas/ClassificationInfo"
          },
          "hint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutOfScopeHint"
              },
              {
                "type": "null"
              }
            ]
          },
          "output": {
            "$ref": "#/components/schemas/DecoderOutput"
          },
          "presentation": {
            "$ref": "#/components/schemas/PresentationInfo"
          }
        },
        "required": [
          "classification",
          "output",
          "brain",
          "presentation",
          "hint"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
...
      "ExperimentReport": {
        "additionalProperties": false,
        "properties": {
          "created_utc": {
            "title": "Created Utc",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "report_version": {
            "default": 1,
            "title": "Report Version",
            "type": "integer"
          },
          "request": {
            "$ref": "#/components/schemas/ExperimentRequest"
          },
          "result": {
            "$ref": "#/components/schemas/ExperimentResult"
          },
          "title_tr": {
            "default": "Laboratuvar deneyi",
            "maxLength": 200,
            "title": "Title Tr",
            "type": "string"
          }
        },
        "required": [
          "created_utc",
          "request",
          "result",
          "provenance"
        ],
        "title": "ExperimentReport",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobState": {
        "enum": [
          "queued",
          "running",
          "done",
          "error"
        ],
        "title": "JobState",
        "type": "string"
      },
      "JobStatus": {
        "additionalProperties": false,
        "description": "Arka planda çalışan Laboratuvar işi. Sonuç, işin türüne göre ilgili alanda döner.",
        "properties": {
          "error_tr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Tr"
          },
          "experiment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExperimentResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "run",
              "pathways",
              "report",
              "replay"
            ],
            "title": "Kind",
            "type": "string"
          },
          "pathways": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PathwayResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "progress": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "replay": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReplayResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExperimentReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "step_tr": {
            "title": "Step Tr",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "state",
          "progress",
          "step_tr"
        ],
        "title": "JobStatus",
        "type": "object"
      },
...
      "PathwayRequest": {
        "additionalProperties": false,
        "properties": {
          "behavior": {
            "$ref": "#/components/schemas/Behavior"
          },
          "experiment": {
            "$ref": "#/components/schemas/ExperimentRequest"
          },
          "max_hops": {
            "default": 4,
            "maximum": 6.0,
            "minimum": 1.0,
            "title": "Max Hops",
            "type": "integer"
          },
          "top_k": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "experiment",
          "behavior"
        ],
        "title": "PathwayRequest",
        "type": "object"
      },
...
      "ReportRequest": {
        "additionalProperties": false,
        "properties": {
          "experiment": {
            "$ref": "#/components/schemas/ExperimentRequest"
          },
          "title_tr": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title Tr"
          }
        },
        "required": [
          "experiment"
        ],
        "title": "ReportRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Drosophila konektomu üzerinde şeffaf beyin simülasyonu API'si",
    "title": "FlyInfo",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/chat": {
      "post": {
        "operationId": "post_chat_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Chat"
      }
    },
...
    "/api/lab/examples/{name}": {
      "get": {
        "operationId": "get_lab_example_api_lab_examples__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Lab Example"
      }
    },
    "/api/lab/jobs/{job_id}": {
      "get": {
        "operationId": "get_lab_job_api_lab_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Lab Job"
      }
    },
...
    "/api/lab/pathways": {
      "post": {
        "operationId": "post_lab_pathways_api_lab_pathways_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PathwayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Lab Pathways"
      }
    },
    "/api/lab/replay": {
      "post": {
        "operationId": "post_lab_replay_api_lab_replay_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Report",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Lab Replay"
      }
    },
    "/api/lab/report": {
      "post": {
        "operationId": "post_lab_report_api_lab_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Lab Report"
      }
    },
    "/api/lab/run": {
      "post": {
        "operationId": "post_lab_run_api_lab_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Lab Run"
      }
    },
    "/api/neurons/search": {
      "get": {
        "operationId": "get_neuron_search_api_neurons_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CellTypeMatch"
                  },
                  "title": "Response Get Neuron Search Api Neurons Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Neuron Search"
      }
    },
    "/api/neurons/{cell_type}": {
      "get": {
        "operationId": "get_neurons_api_neurons__cell_type__get",
        "parameters": [
          {
            "in": "path",
            "name": "cell_type",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "title": "Cell Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CellTypeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Neurons"
      }
    },
...
    "/api/stimulate": {
      "post": {
        "operationId": "post_stimulate_api_stimulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StimulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StimulateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Stimulate"
      }
    }
  }
}